src/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
anthropic==0.40.0
reportlab==4.2.5
shapely==2.0.6
numpy==1.26.4
python-dotenv==1.0.1
python-multipart==0.0.12
